+++ Assets/Terasurware/Classes/Entity_ItemDataBase.cs
@@ -43,7 +43,7 @@ public class Entity_ItemDataBase : ScriptableObject
 		public int watery;
 		public string type;
 		public string subtype;
-		public int girl1_like;
+		public float girl1_like;
 		public int cost_price;
 		public int sell_price;
 		public string topping01;
